KYLIN-752 improve query performance for values are not in dictionary
A quick fix to drop the not-in-dictionary values before generating hbase
scan range. And based on it, we can make a further optimization that as
long as there is at least one andFilter is empty(after we drop the
not-in-dictionary values), we can simply assert that the whole
andFilters is empty to avoid an unnecessary hbase scan. However, this
strategy only works for 'EQ', 'IN'.
For 'LT', 'LTE', 'GT', 'GTE', we have to replace them with the some sort
of closest smaller or bigger values by calling "final public int
getIdFromValue(T value, int roundingFlag)" of "Dictionary" class.